requirements.txt
certbot
dnspython
pyyaml
requests
tencentcloud-sdk-python-common
tencentcloud-sdk-python-dnspod
tldextract
